ram: ast2600: Enable device tree based DDR config
The driver supports specification of the M-PLL via the device tree, and
partially supported using this to configure the SDRAM rate used for
memory training.

This patch finishes the job and exposes the setting through Kconfig.

Note that the following exact values must be used for the
clock-frequency property:

 - 400000000
 - 334000000
 - 200000000
 - 100000000

Any other value will cause the code to default to 334MHz.

Signed-off-by: Joel Stanley <joel@jms.id.au>