configs: Add AST2600 SPL eMMC configuration
U-Boot and environment are to be loaded off the eMMC.
The environment is located at byte offset 0x5000 in the user data
partition, and is sized at 0x10000. The redundant environment
follows it immediately and is the same size. Also select the
partition number and block offset of the U-Boot image; it is in
the first boot partition (eMMC partition 1) at block 128.

Signed-off-by: Eddie James <eajames@linux.ibm.com>
Link: https://lore.kernel.org/r/20201105200920.51226-4-eajames@linux.ibm.com
Signed-off-by: Joel Stanley <joel@jms.id.au>