config: ast2600: Configure common MMC SPL loader
The SPL will load u-boot proper from sector 0x80 (128), which is 64KB
into the device.

The link address changes as this is used as the load address by the
loader. Given the Aspeed loaders are linking u-boot at 0x10000 but
running it from RAM, the u-boot relocation code must be fine with this
setup.

The custom Aspeed loaders are disabled, and so the related configuration
options are removed from the defconfig.

Signed-off-by: Joel Stanley <joel@jms.id.au>