configs: ast2600: Use non-a1 config for openbmc spl emmc
The latest ASPEED SDK created a non-versioned evb configuration.
Previously the openbmc configurations were using the a1 config, this
changes it to use the non-versioned one which will support A2 and
onwards.

There is no difference between the configurations at this stage other
than the Ethernet node in the device tree for the evb.

Signed-off-by: Joel Stanley <joel@jms.id.au>